Matrix initialization and equality comparison.
Also, moved float comparison function to its own module so others can use it (out of tuple). Tried putting it in the assert_eqf64 macro module but then the macro can't be used for some reason. According to the docs, it seems macros need to be in their own external 'crate' but for now this workaround is ok.