Matrix multiplication, transpose, identity. Learned about Copy/Clone
Need to use 'clone' to avoid losing ownership of the matrix when calling a function.

Tuple uses 'copy' because it's made up of primitives (f64), this is like an 'implicit' clone that happens during pass-by-value because it's assumed to be cheap and fast. Dynamically sized types like Vec don't support Copy

Clone is the same, except it requires an explicit call to clone() in order to create the (deep) copy of a struct that implements the Clone trait. It'd be more efficient to either: 1) implement operations to act on references so we don't lose ownership of operands, or 2) hardcode a struct without using Vec, e.g. have a Mat4 class that knows exactly how large it will be and can simply use 16 fields of type f64.

Both of the above are more tedious to program than clone(), though